Make InputSystem  package compile with editor in separate assembly (still many TODOs)
