+++ Packages/com.unity.inputsystem/InputSystem/Actions/Interactions/SlowTapInteraction.cs
@@ -1,7 +1,3 @@
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-#endif
-
 namespace UnityEngine.InputSystem.Interactions
 {
     // Performs the action if the control is pressed, held for at least the set duration
@@ -46,33 +42,4 @@ namespace UnityEngine.InputSystem.Interactions
             m_SlowTapStartTime = 0.0;
         }
     }
-
-    #if UNITY_EDITOR
-    internal class SlowTapInteractionEditor : InputParameterEditor<SlowTapInteraction>
-    {
-        protected override void OnEnable()
-        {
-            m_DurationSetting.Initialize("Min Tap Duration",
-                "Minimum time (in seconds) that a control has to be held for it to register as a slow tap. If the control is released "
-                + "before this time, the slow tap is canceled.",
-                "Default Slow Tap Time",
-                () => target.duration, x => target.duration = x, () => InputSystem.settings.defaultSlowTapTime);
-            m_PressPointSetting.Initialize("Press Point",
-                "The amount of actuation a control requires before being considered pressed. If not set, default to "
-                + "'Default Button Press Point' in the global input settings.",
-                "Default Button Press Point",
-                () => target.pressPoint, v => target.pressPoint = v,
-                () => InputSystem.settings.defaultButtonPressPoint);
-        }
-
-        public override void OnGUI()
-        {
-            m_DurationSetting.OnGUI();
-            m_PressPointSetting.OnGUI();
-        }
-
-        private CustomOrDefaultSetting m_DurationSetting;
-        private CustomOrDefaultSetting m_PressPointSetting;
-    }
-    #endif
 }
