+++ Packages/com.unity.inputsystem/InputSystem/Editor/InputParameterEditor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bd784762c8684645a87c61fdb512cb8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
