+++ Packages/com.unity.inputsystem/InputSystem/Editor/InputParameterEditor/HoldInteractionEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8e196fcc17de45d5910de55e847478d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
