+++ Packages/com.unity.inputsystem/InputSystem/Editor/InputParameterEditor/MultiTapInteractionEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4bdb75cbf1c024f168a55cd15d4df2a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
