+++ Packages/com.unity.inputsystem/InputSystem/Editor/InputParameterEditor/PressInteractionEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f424f79a5892460f95e036911289538
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
