+++ Packages/com.unity.inputsystem/InputSystem/Editor/InputParameterEditor/SlowTapInteractionEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 04c2d92de7d5147dc91ad09dd72b92ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
