+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dccf5361d339d4ecea86436dcdda92fa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
