+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/PlayerInput.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bf4974f1f412446aabbfcd49081e87a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
