+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/Steam.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e3b950e9162e4404b8cdc2c60a6a2f2e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
