+++ Packages/com.unity.inputsystem/InputSystem/Editor/Unity.InputSystem.Editor.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "Unity.InputSystem.Editor",
+    "references": [
+        "GUID:75469ad4d38634e559750d17036d5f7c"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": true,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": []
+}
