+++ Packages/com.unity.inputsystem/InputSystem/Editor/Unity.InputSystem.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f8b00dda8b61f434eb195a416c32d879
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
