+++ Packages/com.unity.inputsystem/Tests/InputSystem/Unity.InputSystem.TestFramework.asmdef
@@ -1,17 +1,21 @@
 {
     "name": "Unity.InputSystem.TestFramework",
     "references": [
-        "GUID:75469ad4d38634e559750d17036d5f7c"
-    ],
-    "optionalUnityReferences": [
-        "TestAssemblies"
+        "GUID:75469ad4d38634e559750d17036d5f7c",
+        "GUID:27619889b8ba8c24980f49ee34dbb44a",
+        "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:f8b00dda8b61f434eb195a416c32d879"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": true,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
     "versionDefines": []
 }
